refactor lib namespace
