+++ lib/spid-rails/onelogin/rubysaml/authrequest.rb
