+++ lib/spid-rails/version.rb
@@ -1,0 +1,7 @@
+module Spid
+  module Rails
+
+    VERSION = '0.1'
+    
+  end
+end
