--- lib/spid_rails/engine.rb
@@ -1,8 +1,0 @@
-require 'onelogin/ruby-saml'
-require 'spid_rails/onelogin/rubysaml/authrequest'
-
-module SpidRails
-  class Engine < ::Rails::Engine
-    isolate_namespace SpidRails
-  end
-end
