--- lib/spid_rails/version.rb
@@ -1,3 +1,0 @@
-module SpidRails
-  VERSION = '0.1'
-end
